+++ xpybuild/internal/__init__.py
@@ -1,0 +1,3 @@
+# need to get this here before we set sys.stdout=None
+import locale, sys
+DEFAULT_PROCESS_ENCODING = sys.stdout.encoding or locale.getpreferredencoding()